Codegen overhaul, v2 upgrade, funds-safety fixes, and release hardening (v5) (#120)

+++ sdk-clients/fusionplus/examples/get_settlement_contract/main.go
@@ -44,7 +44,7 @@ func main() {
 	ctx := context.Background()
 
 	response, err := client.GetSettlementContract(ctx, fusionplus.GetSettlementContractParams{
-		ChainId: float32(constants.EthereumChainId),
+		ChainId: constants.EthereumChainId,
 	})
 	if err != nil {
 		log.Fatalf("failed to get settlement contract: %v", err)
